requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
importlib_resources==6.5.2
pillow==11.1.0
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.1
setuptools==78.1.0
six==1.17.0
